# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RZNU LAB1",
        "description": "Računarstvo zasnovano na uslugama LAB1",
        "contact": {
            "email": "mislav.zabcic@gmail.com"
        },
        "license": {
            "name": "MIT license",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/login": {
            "post": {
                "tags": [
                    "Base"
                ],
                "summary": "Logs user into system",
                "operationId": "authenticate",
                "requestBody": {
                    "$ref": "#/components/requestBodies/Login"
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "headers": {
                            "X-RateLimit-Limit": {
                                "description": "Rate limit for token",
                                "schema": {
                                    "type": "Int64"
                                }
                            },
                            "X-RateLimit-Remaining": {
                                "description": "Rate limit remaining for token",
                                "schema": {
                                    "type": "Int64"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error with body parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/": {
            "get": {
                "tags": [
                    "Base"
                ],
                "summary": "Get welcome message",
                "description": "Get welcome message",
                "operationId": "welcome",
                "responses": {
                    "200": {
                        "description": "Succes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Base"
                ],
                "summary": "Registers user into system",
                "operationId": "register",
                "requestBody": {
                    "$ref": "#/components/requestBodies/Register"
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "headers": {
                            "X-RateLimit-Limit": {
                                "description": "Rate limit for token",
                                "schema": {
                                    "type": "Int64"
                                }
                            },
                            "X-RateLimit-Remaining": {
                                "description": "Rate limit remaining for token",
                                "schema": {
                                    "type": "Int64"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error with body parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all users",
                "description": "Returns array of all registered users",
                "operationId": "getAllUsers",
                "responses": {
                    "200": {
                        "description": "Succes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Creats new user into system",
                "operationId": "create",
                "requestBody": {
                    "$ref": "#/components/requestBodies/Register"
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "headers": {
                            "X-RateLimit-Limit": {
                                "description": "Rate limit for token",
                                "schema": {
                                    "type": "Int64"
                                }
                            },
                            "X-RateLimit-Remaining": {
                                "description": "Rate limit remaining for token",
                                "schema": {
                                    "type": "Int64"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error with body parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/users/me": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get current user",
                "description": "Returns current user",
                "operationId": "getAuthenticatedUser",
                "responses": {
                    "200": {
                        "description": "Succes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
...
        "/users/{userId}/posts": {
            "get": {
                "tags": [
                    "Posts, Users"
                ],
                "summary": "Get posts by user id",
                "description": "Returns posts that where created by user",
                "operationId": "getPostByUserId",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID of a user for posts to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Post"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "AuthResponse": {
                "title": "AuthResponse",
                "description": "Register and login api response",
                "properties": {
                    "user": {
                        "title": "User",
                        "description": "User authentificated",
                        "$ref": "#/components/schemas/User"
                    },
                    "token": {
                        "title": "Token",
                        "description": "JWT token",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Error": {
                "title": "Error",
                "description": "Error api response",
                "properties": {
                    "error": {
                        "title": "Error",
                        "description": "Error message",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Message": {
                "title": "Message",
                "description": "Message api response",
                "properties": {
                    "message": {
                        "title": "Message",
                        "description": "Message message",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "requestBodies": {
...
            "Login": {
                "description": "Login object that needs to be authentificated",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Login"
                        }
                    }
                }
            },
            "Register": {
                "description": "Register object for registration and creating new user",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Register"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "Authorization": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}